% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_permutation4}
\alias{generate_permutation4}
\title{return a permutation of numbers with ununiform probabilities of being chosen.}
\usage{
generate_permutation4(permutation_size, ununiform_probabilities0)
}
\arguments{
\item{permutation_size}{size of the permutation}

\item{ununiform_probabilities0}{relative probabilities of being chosen}
}
\description{
return a permutation of numbers with ununiform probabilities of being chosen.
}
