% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IESTR_user_functions.R
\name{IESTR_process}
\alias{IESTR_process}
\title{Whole IESTR process}
\usage{
IESTR_process(
  suitability_maps,
  presence_map,
  migr_spe,
  cost_map,
  threshold,
  confidence,
  npop,
  nsur,
  ngen
)
}
\arguments{
\item{suitability_maps}{list of matrix of suitability for each timestep.}

\item{presence_map}{sparse matrix of presence of the species.}

\item{migr_spe}{matrix of probability of spread in neighbour cells at each time step.}

\item{cost_map}{Cost of introduction of the species in each cell.}

\item{threshold}{Number of sites with presence is wanted at the end of the study period}

\item{confidence}{Probability that the threshold is reached is wanted (must be <1)}

\item{npop}{Genetic algorithm : size of the initial population}

\item{nsur}{Genetic algorithm : size of the survival population at each generation}

\item{ngen}{Genetic algorithm : number of generation}
}
\value{
the transition matrices & the choices of introductions
}
\description{
Whole IESTR process
}
