% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_random_choice_non_uniform}
\alias{index_random_choice_non_uniform}
\title{return random value from ununiform probability set.}
\usage{
index_random_choice_non_uniform(ununiform_probabilities0)
}
\arguments{
\item{ununiform_probabilities0}{Rcpp vector of probabilities for each indices value}
}
\description{
return random value from ununiform probability set.
}
