% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proba_matrix_mult3}
\alias{proba_matrix_mult3}
\title{return the special matrix product between two probability matrix.}
\usage{
proba_matrix_mult3(A, B)
}
\arguments{
\item{A}{first probability matrix}

\item{B}{second probability matrix}
}
\description{
return the special matrix product between two probability matrix.
}
