% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomfunc3}
\alias{randomfunc3}
\title{get a random value between 0 & (j-1)}
\usage{
randomfunc3(j)
}
\arguments{
\item{j}{limit-1 of random value}
}
\description{
get a random value between 0 & (j-1)
}
