% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_algorithm_opt}
\alias{rcpp_algorithm_opt}
\title{Genetic algorithm used to optimised choices of introduction under the constraint of reaching a dertain number of 
presence at the end of the study period (threshold) with a minimal probability (confidence) while miimizing the cost
of all the introductions}
\usage{
rcpp_algorithm_opt(
  pheromons,
  viablesTriplets,
  population0,
  costMatrix,
  currentPresenceMatrix,
  transitionmatrices,
  globalSuitableSites,
  viablesValues,
  threshold,
  confidence,
  npop,
  nsur,
  ngen,
  nbtoplant
)
}
\arguments{
\item{Rcpp::NumericVector}{pheromons weights of relative importance of each site&time pair.}

\item{Rcpp::NumericMatrix}{population0 An initial population of set of site&time choices of introduction}

\item{std::vector<Eigen::SparseMatrix<double>>}{transitionmatriceslist of transition matrices from any timestep to the last timestep.}

\item{Eigen::SparseMatrix<double>}{viablesValues viables values from transitions matrices from viable sites.}

\item{double}{confidence minimal probability to reach the threshold accepted by the optimisation algorithm}

\item{int}{ngen genetic algorithm : number of generation}
}
\value{
an optimised population matrix of set of site&time choices of introduction.
}
\description{
Genetic algorithm used to optimised choices of introduction under the constraint of reaching a dertain number of 
presence at the end of the study period (threshold) with a minimal probability (confidence) while miimizing the cost
of all the introductions
}
