% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_algorithm_opt2}
\alias{rcpp_algorithm_opt2}
\title{rcpp_algorithm_opt2}
\usage{
rcpp_algorithm_opt2(
  pheromons,
  viablesTriplets,
  population0,
  costMatrix,
  currentPresenceMatrix,
  transitionmatrices,
  globalSuitableSites,
  viablesValues,
  threshold,
  confidence,
  npop,
  nsur,
  ngen,
  nbtoplant
)
}
\arguments{
\item{pheromons}{(Rcpp::NumericVector) weights of relative importance of each site&time pair.}

\item{viablesTriplets}{(Rcpp::NumericMatrix) For each viable site : index, position on map, time, and cost.}

\item{population0}{(Rcpp::NumericMatrix) An initial population of set of site&time choices of introduction}

\item{costMatrix}{(Eigen::SparseMatrix<double>) matrix of costs of introduction, sparse.}

\item{currentPresenceMatrix}{(Eigen::SparseMatrix<double>) matrix of current presence of the species, sparse.}

\item{transitionmatrices}{(std::vector<Eigen::SparseMatrix<double>>) list of transition matrices from any timestep to the last timestep.}

\item{globalSuitableSites}{(Eigen::SparseMatrix<double>) matrix of viable sites to use for the optimiation part.}

\item{viablesValues}{(Eigen::SparseMatrix<double>) viables values from transitions matrices from viable sites.}

\item{threshold}{(int) number of presence site to reach at the end of the studied period, with the introductions}

\item{confidence}{(double) minimal probability to reach the threshold accepted by the optimisation algorithm}

\item{npop}{(int) genetic algorithm : size of the population}

\item{nsur}{(int) genetic algorithm : size of the surviving population}

\item{ngen}{(int) genetic algorithm : number of generations}

\item{nbtoplant}{evaluated maximum number of introduction necessary to satisfy the constraints of final presence.}
}
\value{
an optimised population matrix of set of site&time choices of introduction.
}
\description{
Genetic algorithm used to optimised choices of introduction under the constraint of reaching a dertain number of 
presence at the end of the study period (threshold) with a minimal probability (confidence) while miimizing the cost
of all the introductions
}
