% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_eval_current_prob}
\alias{rcpp_eval_current_prob}
\title{rcpp_eval_current_prob}
\usage{
rcpp_eval_current_prob(
  threshold,
  currentPresenceMatrix,
  transitionmatrices,
  globalSuitableSites
)
}
\arguments{
\item{threshold}{(int) number of site with presence we want to obtain.}

\item{currentPresenceMatrix}{(Eigen::SparseMatrix<double>) matrix of presence of the species.}

\item{transitionmatrices}{(std::vector<Eigen::SparseMatrix<double>>) list of transition matrices from any timestep to the last timestep.}

\item{globalSuitableSites}{(Eigen::SparseMatrix<double>) support matrix containing index of each sites}
}
\value{
for each viable site : index, position on map, time, and cost.
}
\description{
Evaluate how would evolve the current present species without introduction.
}
