% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_generate_population}
\alias{rcpp_generate_population}
\title{Genrate initial population for the genetic algorithm.}
\usage{
rcpp_generate_population(pheromons, globalSuitableSites, npop, nbtoplant)
}
\arguments{
\item{Rcpp::NumericVector}{pheromons weights of relative importance of each site&time pair.}

\item{Eigen::SparseMatrix<double>}{globalSuitableSites support matrix containing index of each sites}

\item{int}{nbtoplant pre-evaluate number of introduction necessary (overestimated, usually)}
}
\value{
an initial population of set of site&time choices of introduction
}
\description{
Genrate initial population for the genetic algorithm.
}
