% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_global_suitable_coordinates}
\alias{rcpp_global_suitable_coordinates}
\title{Calculate the coordinates of the suitable sites.}
\usage{
rcpp_global_suitable_coordinates(globalSuitableSites)
}
\arguments{
\item{Eigen::SparseMatrix<double>}{globalSuitableSites}
}
\value{
coordinates of the suitable sites as a matrix
}
\description{
Calculate the coordinates of the suitable sites.
}
