% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_result_to_choice}
\alias{rcpp_result_to_choice}
\title{Rewriting the final population given by the genetic algorithm with coordinates and times of introduction}
\usage{
rcpp_result_to_choice(lastPopulation, viablesTriplets)
}
\arguments{
\item{Rcpp::NumericMatrix}{viablesTriplets a matrix of information about each viable site.}
}
\value{
the choices of introduction in coordinates and timesteps.
}
\description{
Rewriting the final population given by the genetic algorithm with coordinates and times of introduction
}
