% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_spread_matrix}
\alias{rcpp_spread_matrix}
\title{rcpp_spread_matrix}
\usage{
rcpp_spread_matrix(
  globalSuitableSites,
  globalSuitableCoordinates,
  migrationKernel
)
}
\arguments{
\item{globalSuitableSites}{(Eigen::SparseMatrix<double>) support matrix containing index of each sites}

\item{globalSuitableCoordinates}{(Rcpp::NumericMatrix) support matrix coordinates of each sites}

\item{migrationKernel}{(Rcpp::NumericMatrix) matrix modelling spread properties of the species.}
}
\value{
Spread matrix used to model spread behavior of the species without suitability
}
\description{
Calculate the spread_matrix.
}
