% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_transition_matrix}
\alias{rcpp_transition_matrix}
\title{rcpp_transition_matrix}
\usage{
rcpp_transition_matrix(localtransitionmatrices)
}
\arguments{
\item{localtransitionmatrices}{(std::list<Eigen::SparseMatrix<double>>) list of matrix used to model spread behavior of the species with suitability, between two adjacent timestep}
}
\value{
list of transition matrices from any timestep to the last timestep.
}
\description{
Calculate the transition matrices from the local transition matrices
}
