% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_viable_values}
\alias{rcpp_viable_values}
\title{rcpp_viable_values}
\usage{
rcpp_viable_values(
  viablesTriplets,
  viableSites,
  globalSuitableSites,
  transitionmatrices
)
}
\arguments{
\item{viablesTriplets}{(Rcpp::NumericMatrix) For each viable site : index, position on map, time, and cost.}

\item{viableSites}{(Eigen::SparseMatrix<double>) boolean sparse matrix, each row correspond to a time, each column to a site. (1=viable; 0=not viable)}

\item{globalSuitableSites}{(Eigen::SparseMatrix<double> ) support matrix containing index of each sites}

\item{transitionmatrices}{(std::vector<Eigen::SparseMatrix<double>>) list of transition matrices from any timestep to the last timestep.}
}
\value{
for each viable site : index, position on map, time, and cost.
}
\description{
Summing up the effect on the final state of the system of each viable site&time pair.
}
