% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_eval_current_prob}
\alias{rcpp_eval_current_prob}
\title{Evaluate how would evolve the current present species without introduction.}
\usage{
rcpp_eval_current_prob(
  threshold,
  currentPresenceMatrix,
  transitionmatrices,
  globalSuitableSites
)
}
\arguments{
\item{threshold}{number of site with presence we want to obtain.}

\item{std::vector<Eigen::SparseMatrix<double>>}{transitionmatrices list of transition matrices from any timestep to the last timestep.}

\item{Eigen::SparseMatrix<double>}{globalSuitableSites support matrix containing index of each sites}
}
\value{
for each viable site : index, position on map, time, and cost.
}
\description{
Evaluate how would evolve the current present species without introduction.
}
