% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_generate_population}
\alias{rcpp_generate_population}
\title{rcpp_generate_population}
\usage{
rcpp_generate_population(pheromons, globalSuitableSites, npop, nbtoplant)
}
\arguments{
\item{pheromons}{(Rcpp::NumericVector) weights of relative importance of each site&time pair.}

\item{globalSuitableSites}{(Eigen::SparseMatrix<double>) support matrix containing index of each sites}

\item{npop}{(int) number of individuals in the initial population}

\item{nbtoplant}{(int) pre-evaluate number of introduction necessary (overestimated, usually)}
}
\value{
an initial population of set of site&time choices of introduction
}
\description{
Generate initial population for the genetic algorithm.
}
