% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_global_suitable_coordinates}
\alias{rcpp_global_suitable_coordinates}
\title{rcpp_global_suitable_coordinates.}
\usage{
rcpp_global_suitable_coordinates(globalSuitableSites)
}
\arguments{
\item{globalSuitableSites}{(Eigen::SparseMatrix<double>) support matrix containing index of each sites}
}
\value{
coordinates of the suitable sites as a matrix
}
\description{
Calculate the coordinates of the suitable sites.
}
