% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_global_suitable_sites}
\alias{rcpp_global_suitable_sites}
\title{rcpp_global_suitable_sites}
\usage{
rcpp_global_suitable_sites(consecutiveSuitabilityMatrix)
}
\arguments{
\item{consecutiveSuitabilityMatrix}{(std::list<Eigen::SparseMatrix<double>>) list of matrix of suitability for each timestep.}
}
\value{
suitable sites as a sparse matrix
}
\description{
Establishes from the consecutiveSuitabilityMatrix what are the
study sites in the map and attribute them an index used in 
other functions.
Warning : the indices start at 1.
}
