% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_local_transition_matrix}
\alias{rcpp_local_transition_matrix}
\title{Calculate the local transition matrices from the spread matrix and suitability matrices}
\usage{
rcpp_local_transition_matrix(
  consecutiveSuitabilityMatrix,
  localTransitionMatrix,
  globalSuitableCoordinates
)
}
\arguments{
\item{std::list<Eigen::SparseMatrix<double>>}{consecutiveSuitabilityMatrix list of suitability matrix}

\item{Eigen::SparseMatrix<double>}{localTransitionMatrix matrix used to model spread behavior of the species without suitability}

\item{Rcpp::NumericMatrix}{globalSuitableCoordinates support matrix coordinates of each sites}
}
\value{
list of matrix used to model spread behavior of the species with suitability, between two adjacent timestep
}
\description{
Calculate the local transition matrices from the spread matrix and suitability matrices
}
