% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_local_transition_matrix}
\alias{rcpp_local_transition_matrix}
\title{Multiplies two doubles}
\usage{
rcpp_local_transition_matrix(
  consecutiveSuitabilityMatrix,
  localTransitionMatrix,
  globalSuitableCoordinates
)
}
\arguments{
\item{std::list<Eigen::SparseMatrix<double>>}{consecutiveSuitabilityMatrix}

\item{Eigen::SparseMatrix<double>}{localTransitionMatrix}

\item{Rcpp::NumericMatrix}{globalSuitableCoordinates}
}
\value{
Product of v1 and v2
}
\description{
Multiplies two doubles
}
