% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_viable_triplets}
\alias{rcpp_viable_triplets}
\title{Returns matrix summing up information of each viable site&time pair.}
\usage{
rcpp_viable_triplets(
  viableSites,
  transitionmatrices,
  globalSuitableCoordinates,
  globalSuitableSites,
  costMatrix
)
}
\arguments{
\item{std::list<Eigen::SparseMatrix<double>>}{transitionmatrices list of transition matrices from any timestep to the last timestep.}

\item{Rcpp::NumericMatrix}{globalSuitableCoordinates support matrix coordinates of each sites}

\item{Eigen::SparseMatrix<double>}{costMatrix matrix of cost of introduction}
}
\value{
For each viable site : index, position on map, time, and cost.
}
\description{
Returns matrix summing up information of each viable site&time pair.
}
