% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_viable_values}
\alias{rcpp_viable_values}
\title{Summing up the effect on the final state of the system of each viable site&time pair.}
\usage{
rcpp_viable_values(
  viablesTriplets,
  viableSites,
  globalSuitableSites,
  colonisationMatrices
)
}
\arguments{
\item{Rcpp::NumericMatrix}{viablesTriplets For each viable site : index, position on map, time, and cost.}

\item{Eigen::SparseMatrix<double>}{globalSuitableSites support matrix containing index of each sites}

\item{std::vector<Eigen::SparseMatrix<double>>}{colonisationMatrices list of transition matrices from any timestep to the last timestep.}
}
\value{
for each viable site : index, position on map, time, and cost.
}
\description{
Summing up the effect on the final state of the system of each viable site&time pair.
}
